Some changes to mapsystem

+++ ModManager/MapSystem/MapInstaller.cs
@@ -3,8 +3,10 @@ using ModManager.AddonInstallerSystem;
 using ModManager.AddonSystem;
 using ModManager.ModIoSystem;
 using ModManager.PersistenceSystem;
+using System;
 using System.Collections.Generic;
 using System.IO;
+using System.IO.Compression;
 using System.Linq;
 using File = Modio.Models.File;
 
@@ -18,11 +20,14 @@ namespace ModManager.MapSystem
 
         private readonly ExtractorService _extractor;
 
+        private readonly MapManifestFinder _mapManifestFinder;
+
         public MapInstaller()
         {
             _persistenceService = PersistenceService.Instance;
             _installedAddonRepository = InstalledAddonRepository.Instance;
             _extractor = ExtractorService.Instance;
+            _mapManifestFinder = new MapManifestFinder();
         }
 
         public bool Install(Mod mod, string zipLocation)
@@ -31,17 +36,20 @@ namespace ModManager.MapSystem
             {
                 return false;
             }
+            var zipFile = ZipFile.OpenRead(zipLocation);
+            var timberFileName = zipFile.Entries
+                                    .Where(x => x.Name.Contains(".timber"))
+                                    .Single()
+                                    .Name;
+            zipFile.Dispose();
             string installLocation = _extractor.ExtractMap(zipLocation, mod);
 
             // TODO: manifest handling is probably wrong atm
-            string manifestPath = Path.Combine(installLocation, Manifest.FileName);
-            List<MapManifest> manifests = new();
-            if (System.IO.File.Exists(manifestPath))
-            {
-                manifests = _persistenceService.LoadObject<List<MapManifest>>(manifestPath, false);
-            }
-            var manifest = new MapManifest(mod, mod.Modfile, installLocation);
+            var manifest = new MapManifest(mod, mod.Modfile, installLocation, timberFileName);
+
+            var manifests = _mapManifestFinder.Find().Select(a => (MapManifest)a).ToList();
             manifests.Add(manifest);
+
             string mapManifestPath = Path.Combine(installLocation, Manifest.FileName);
             _persistenceService.SaveObject(manifests, mapManifestPath);
             _installedAddonRepository.Add(manifest);
@@ -51,7 +59,22 @@ namespace ModManager.MapSystem
 
         public bool Uninstall(Manifest manifest)
         {
+            Console.WriteLine($"inside mapinstaller");
+            if (manifest is not MapManifest)
+            {
+                Console.WriteLine($"\"{manifest.ModName} is not a map.");
+                return false;
+            }
+
+            string manifestPath = Path.Combine(Paths.Maps, Manifest.FileName);
+            var manifests = _mapManifestFinder.Find().Select(a => (MapManifest)a).ToList();
+            manifests.Remove(manifests.Where(x => x.ModId == manifest.ModId).SingleOrDefault());
+            _persistenceService.SaveObject(manifests, manifestPath);
+
             _installedAddonRepository.Remove(manifest.ModId);
+
+            var mapFullPath = Path.Combine(Paths.Maps, ((MapManifest)manifest).MapFileName);
+            System.IO.File.Delete(mapFullPath);
 
             return true;
         }
